# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WritingPrompt {
  id: number;
  text: string;
  category: string;
}

export interface JournalEntry {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  content: string;
  /** @nullable */
  promptId?: number | null;
  /** @nullable */
  promptText?: string | null;
  /**
     * Mood rating 1-5 linked to this entry
     * @nullable
     */
  moodRating?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface EntryList {
  entries: JournalEntry[];
  total: number;
}

export interface EntryInput {
  /** @minLength 1 */
  content: string;
  /** Date in YYYY-MM-DD format (defaults to today) */
  date?: string;
  /** @nullable */
  promptId?: number | null;
  /**
     * 1=awful, 2=bad, 3=okay, 4=good, 5=great
     * @nullable
     */
  moodRating?: number | null;
}

export interface EntryUpdate {
  /** @minLength 1 */
  content?: string;
  /** @nullable */
  moodRating?: number | null;
}

export interface EntryStats {
  totalEntries: number;
  /** Days in a row with an entry */
  currentStreak: number;
  longestStreak: number;
  thisWeekCount: number;
  thisMonthCount: number;
}

export type MoodLogMood = typeof MoodLogMood[keyof typeof MoodLogMood];


export const MoodLogMood = {
  awful: 'awful',
  bad: 'bad',
  okay: 'okay',
  good: 'good',
  great: 'great',
} as const;

export interface MoodLog {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  mood: MoodLogMood;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export type MoodInputMood = typeof MoodInputMood[keyof typeof MoodInputMood];


export const MoodInputMood = {
  awful: 'awful',
  bad: 'bad',
  okay: 'okay',
  good: 'good',
  great: 'great',
} as const;

export interface MoodInput {
  mood: MoodInputMood;
  /** @nullable */
  note?: string | null;
  /** Date in YYYY-MM-DD format (defaults to today) */
  date?: string;
}

export type MoodCalendarDayMood = typeof MoodCalendarDayMood[keyof typeof MoodCalendarDayMood];


export const MoodCalendarDayMood = {
  awful: 'awful',
  bad: 'bad',
  okay: 'okay',
  good: 'good',
  great: 'great',
} as const;

export interface MoodCalendarDay {
  date: string;
  mood: MoodCalendarDayMood;
  /** @nullable */
  note?: string | null;
}

export interface MoodBreakdown {
  awful: number;
  bad: number;
  okay: number;
  good: number;
  great: number;
}

export interface MoodStats {
  totalLogs: number;
  /** Days in a row with a mood log */
  currentStreak: number;
  /** Average mood 1-5 */
  averageMoodScore: number;
  moodBreakdown: MoodBreakdown;
  /** Last 7 days of moods */
  recentTrend: MoodCalendarDay[];
}

export interface Dashboard {
  todayPrompt: WritingPrompt;
  recentEntries: JournalEntry[];
  todayMood: MoodLog | null;
  entryStats: EntryStats;
  moodStats: MoodStats;
}

export type ListEntriesParams = {
limit?: number;
offset?: number;
/**
 * Filter by month YYYY-MM
 */
month?: string;
};

export type ListMoodsParams = {
limit?: number;
offset?: number;
};

export type GetMoodCalendarParams = {
/**
 * Month in YYYY-MM format
 */
month: string;
};