# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryStats } from './entryStats';
import type { JournalEntry } from './journalEntry';
import type { MoodLog } from './moodLog';
import type { MoodStats } from './moodStats';
import type { WritingPrompt } from './writingPrompt';

export interface Dashboard {
  todayPrompt: WritingPrompt;
  recentEntries: JournalEntry[];
  todayMood: MoodLog | null;
  entryStats: EntryStats;
  moodStats: MoodStats;
}