# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface EntryInput {
  /** @minLength 1 */
  content: string;
  /** Date in YYYY-MM-DD format (defaults to today) */
  date?: string;
  /** @nullable */
  promptId?: number | null;
  /**
     * 1=awful, 2=bad, 3=okay, 4=good, 5=great
     * @nullable
     */
  moodRating?: number | null;
}