# lib/api-zod/src/generated/types/entryList.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntry } from './journalEntry';

export interface EntryList {
  entries: JournalEntry[];
  total: number;
}