# lib/api-zod/src/generated/types/entryStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface EntryStats {
  totalEntries: number;
  /** Days in a row with an entry */
  currentStreak: number;
  longestStreak: number;
  thisWeekCount: number;
  thisMonthCount: number;
}