# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntry {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  content: string;
  /** @nullable */
  promptId?: number | null;
  /** @nullable */
  promptText?: string | null;
  /**
     * Mood rating 1-5 linked to this entry
     * @nullable
     */
  moodRating?: number | null;
  createdAt: string;
  updatedAt: string;
}