# lib/api-zod/src/generated/types/moodCalendarDay.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { MoodCalendarDayMood } from './moodCalendarDayMood';

export interface MoodCalendarDay {
  date: string;
  mood: MoodCalendarDayMood;
  /** @nullable */
  note?: string | null;
}