# lib/api-zod/src/generated/types/moodInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { MoodInputMood } from './moodInputMood';

export interface MoodInput {
  mood: MoodInputMood;
  /** @nullable */
  note?: string | null;
  /** Date in YYYY-MM-DD format (defaults to today) */
  date?: string;
}