# lib/api-zod/src/generated/types/moodStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { MoodBreakdown } from './moodBreakdown';
import type { MoodCalendarDay } from './moodCalendarDay';

export interface MoodStats {
  totalLogs: number;
  /** Days in a row with a mood log */
  currentStreak: number;
  /** Average mood 1-5 */
  averageMoodScore: number;
  moodBreakdown: MoodBreakdown;
  /** Last 7 days of moods */
  recentTrend: MoodCalendarDay[];
}