# lib/api-zod/src/generated/types/writingPrompt.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mental Wellness Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface WritingPrompt {
  id: number;
  text: string;
  category: string;
}